# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API APP",
        "description": "This is API for Nail App",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/store/mail-confirm-update-email": {
            "post": {
                "tags": [
                    "store"
                ],
                "summary": "Send email confirm update new email",
                "operationId": "mail_confirm_update_email",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "abc@gmail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User login",
                "operationId": "user_login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone": "0967258205",
                                    "password": "123456789"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User register",
                "operationId": "user_register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "file"
                                    },
                                    "user_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/user/password": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "User password change ",
                "operationId": "user_password_change",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "old_password": "0967258205",
                                    "password": "123456789",
                                    "confirm_password": "123456789"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/user/me": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Send token on server and return data user",
                "operationId": "user_detail",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "user logout",
                "operationId": "user_logout",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/user/mail-reset-password": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "send email reset password user",
                "operationId": "user_mail_reset_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "abc@gmail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/user/update-user-image": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "update image user",
                "operationId": "update_image_user",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "file"
                                    },
                                    "_method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Successful Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized "
                    },
                    "403": {
                        "description": "Forbidden "
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Bearer Token.",
                "in": "header",
                "scheme": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}